+++ spec/aspect_spec.lua
@@ -1,4 +1,5 @@
 package.path = "./src/?.lua;" .. package.path
+local config = require("aspect.config")
 local aspect = require("aspect.template")
 local tokenizer = require("aspect.tokenizer")
 local astree = require("aspect.ast")
@@ -47,6 +48,55 @@ local large = {}
 for i=1, 1000 do
     large[i] = 10 + i
 end
+
+config.date.months_locale.ru = {
+    [1]  = {"Янв", "Январь"},
+    [2]  = {"Фев", "Февраль"},
+    [3]  = {"Мар", "Март"},
+    [4]  = {"Апр", "Апрель"},
+    [5]  = {"Май", "Май"},
+    [6]  = {"Июн", "Июнь"},
+    [7]  = {"Июл", "Июль"},
+    [8]  = {"Авг", "Август"},
+    [9]  = {"Сен", "Сентябрь"},
+    [10] = {"Окт", "Октябрь"},
+    [11] = {"Ноя", "Ноябрь"},
+    [12] = {"Дек", "Декабрь"},
+}
+config.date.months["янв"] = 1
+config.date.months["фев"] = 2
+config.date.months["мар"] = 3
+config.date.months["апр"] = 4
+config.date.months["май"] = 5
+config.date.months["июн"] = 6
+config.date.months["июл"] = 7
+config.date.months["авг"] = 8
+config.date.months["сен"] = 9
+config.date.months["окт"] = 10
+config.date.months["ноя"] = 11
+config.date.months["дек"] = 12
+
+config.date.months["январь"]  = 1
+config.date.months["февраль"] = 2
+config.date.months["март"]    = 3
+config.date.months["апрель"]  = 4
+config.date.months["май"]     = 5
+config.date.months["июнь"]    = 6
+config.date.months["июль"]    = 7
+config.date.months["август"]  = 8
+config.date.months["сенябрь"] = 9
+config.date.months["октябрь"] = 10
+config.date.months["ноябрь"]  = 11
+config.date.months["декабрь"] = 12
+config.date.week_locale.ru = {
+    [1] = {"Пн", "Понедельник"},
+    [2] = {"Вт", "Вторник"},
+    [3] = {"Ср", "Среда"},
+    [4] = {"Чт", "Четверг"},
+    [5] = {"Пт", "Пятница"},
+    [6] = {"Сб", "Суббота"},
+    [7] = {"Вс", "Воскресение"},
+}
 
 local vars = {
     integer_0 = 0,
@@ -1076,7 +1126,10 @@ end)
 
 describe("Testing date.", function ()
     local local_offset = date.local_offset
-    local custom_offset = 16 * 60 * 60 -- use not existing timezone: +16 UTC
+    local local_offset_name = date.get_timezone(local_offset, "", false)
+    local offset1600 = 16 * 60 * 60 -- use not existing timezone: +16 UTC
+    local offset400 = 4 * 60 * 60
+
     local dates = {
         -- zero zone
         {"2009-02-13 23:31:30 UTC+00",       1234567890, 0},
@@ -1090,15 +1143,17 @@ describe("Testing date.", function ()
         {"2009-02-13T23:31:30",               1234567890 - local_offset, local_offset},
         {"Fri, 13 Feb 2009 23:31:30",         1234567890 - local_offset, local_offset},
         {"Friday, 13-February-2009 23:31:30", 1234567890 - local_offset, local_offset},
+        {"Пт, 13 Фев 2009 23:31:30",          1234567890 - local_offset, local_offset},
+        {"Пятница, 13-Февраль-2009 23:31:30", 1234567890 - local_offset, local_offset},
         {"2009-02-13T23:31:30",               1234567890 - local_offset, local_offset},
-        {"Fri, 13 Feb 2009 23:31:30",         1234567890 - local_offset, local_offset},
+        {"Пт, 13 Февраль 2009 23:31:30",      1234567890 - local_offset, local_offset},
         -- custom zone
-        {"2009-02-13 23:31:30 UTC+16",                   1234567890 - custom_offset, custom_offset},
-        {"2009-02-13T23:31:30+16:00",                    1234567890 - custom_offset, custom_offset},
-        {"Fri, 13 Feb 2009 23:31:30 +1600",              1234567890 - custom_offset, custom_offset},
-        {"Friday, 13-February-2009 23:31:30 UTC +16:00", 1234567890 - custom_offset, custom_offset},
-        {"2009-02-13T23:31:30+16:00",                    1234567890 - custom_offset, custom_offset},
-        {"Fri, 13 Feb 2009 23:31:30 GMT+1600",           1234567890 - custom_offset, custom_offset},
+        {"2009-02-13 23:31:30 UTC+16",                   1234567890 - offset1600, offset1600},
+        {"2009-02-13T23:31:30+16:00",                    1234567890 - offset1600, offset1600},
+        {"Fri, 13 Feb 2009 23:31:30 +1600",              1234567890 - offset1600, offset1600},
+        {"Friday, 13-February-2009 23:31:30 UTC +16:00", 1234567890 - offset1600, offset1600},
+        {"2009-02-13T23:31:30+16:00",                    1234567890 - offset1600, offset1600},
+        {"Fri, 13 Feb 2009 23:31:30 GMT+1600",           1234567890 - offset1600, offset1600},
         -- edge cases
         {"00:00:01 UTC", 1, 0},
         {1, 1, 0},
@@ -1111,6 +1166,7 @@ describe("Testing date.", function ()
                     .. dt.time .. " offset " .. dt.offset .. "; info: " .. dump_table(dt.info))
             assert.is.equals(d[3], dt.offset)
         end)
+
     end
     local date1 = "2009-02-13 23:31:30 UTC+00"
     local date2 = "2009-02-13T23:31:35+00:00"
@@ -1161,20 +1217,37 @@ describe("Testing date.", function ()
         {
             date = "2009-02-13 23:31:30 UTC+16",
             format = {
-                {"%F %T", os.date("%F %T", 1234567890 - custom_offset + local_offset)},
-                {"!%F %T", "2009-02-13 23:31:30"},
-                {"%F %T UTC%z", os.date("%F %T UTC%z", 1234567890)},
-                {"!%F %T UTC%z", "2009-02-13 23:31:30 UTC+00"},
+                { "%F %T UTC%Z", "2009-02-13 23:31:30 UTC+16", offset1600 }, -- UTC+16
+                { "!%F %T UTC%Z", "2009-02-13 07:31:30 UTC", offset1600 }, -- UTC+0 time
+                {"%F %T UTC%Z", "2009-02-13 07:31:30 UTC", 0}, -- UTC+0 time
+                {"!%F %T UTC%Z", "2009-02-13 07:31:30 UTC", 0}, -- UTC+0 time
+                { "%F %T UTC%Z", "2009-02-13 11:31:30 UTC+4", offset400 }, -- UTC+4 time
+                { "!%F %T UTC%Z", "2009-02-13 07:31:30 UTC", offset400 }, -- UTC+4 time
+                {"%F %T UTC"..local_offset_name, os.date("!%F %T UTC%z", 1234567890 + local_offset - offset1600), nil},
+                {"!%F %T UTC%Z", os.date("!%F %T UTC", 1234567890 - offset1600), nil},
             }
-        }
+        },
+        -- locale
+        {
+            date = "2009-02-13 23:31:30 UTC+00",
+            format = {
+                {"%a, %d %b", "Пт, 13 Фев", 0, 'ru'},
+                {"%A, %d %B", "Пятница, 13 Февраль", 0, 'ru'},
+                {"%a, %d %b", "Fri, 13 Feb", 0, 'en'},
+                {"%A, %d %B", "Friday, 13 February", 0, 'en'},
+                {"%a, %d %b", "Fri, 13 Feb", 0},
+                {"%A, %d %B", "Friday, 13 February", 0},
+            }
+        },
     }
     for _, v1 in ipairs(formated) do
         for _, v2 in ipairs(v1.format) do
-            --it("Formatting date '" .. v1.date .. "' as '" .. v2[1] .. "'", function ()
-            --    local d = date.new(v1.date)
-            --    print("d.time: " .. d.time)
-            --    assert.is.equals(v2[2], d:format(v2[1]))
-            --end)
+            it("Formatting date '" .. v1.date .. "' as '" .. v2[1]
+                .. "' with offset " .. tostring(v2[2])
+                .. " and locale " ..tostring(v2[4]), function ()
+                local d = date.new(v1.date)
+                assert.is.equals(v2[2], d:format(v2[1], v2[3], v2[4]))
+            end)
         end
     end
 end)