+++ changelog.md
@@ -11,6 +11,11 @@ Develop
 - add compiler contexts
 - improve template inheritance algorithm
 
+Date:
+ 
+- UTC time offset now in seconds (instead of minutes) 
+- When formatting a date, if a time zone is not specified, then formatting will occur for the local time zone
+
 1.14
 ----